Remove PlainTextResponse response_class (#544)
https://nvidia.slack.com/archives/C08TG7CLEGY/p1766191655660079

Initially in #290 , the `response_class=PlainTextResponse` was added to
the `/global_config_dict_yaml` endpoint of the HeadServer as an attempt
to debug parsing server info for the `ng_status` command. This lead to a
parsing error in `load_from_global_config`. This command now uses it's
own separate endpoint `server_instances`, so this needs to be removed.

Signed-off-by: Frankie Siino <fsiino@nvidia.com>